requirements.txt
# GenAI Legal Tool - Python Dependencies
# Core Framework Dependencies
streamlit==1.49.1
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.3
pydantic==2.11.8
pydantic_core==2.33.2

# PDF Processing Libraries
pdfplumber==0.11.7
PyMuPDF==1.26.4
pypdfium2==4.30.0
pdfminer.six==20250506

# OCR and Image Processing
pytesseract==0.3.13
pillow==11.3.0

# Document Processing
python-docx==1.2.0
lxml==6.0.1
openpyxl==3.1.2

# AI and Machine Learning
openai==1.107.2
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
faiss-cpu==1.12.0
numpy==2.3.3
pandas==2.3.2

# Data Visualization (for Streamlit)
altair==5.5.0
pydeck==0.9.1

# HTTP and API
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
h11==0.16.0

# Utilities
click==8.2.1
colorama==0.4.6
tqdm==4.67.1
tenacity==9.1.2
watchdog==6.0.0
toml==0.10.2
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Security and Cryptography
cryptography==45.0.7
certifi==2025.8.3
cffi==2.0.0
pycparser==2.23

# JSON and Data Processing
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.36.2
rpds-py==0.27.1

# System and Environment
distro==1.9.0
anyio==4.10.0
sniffio==1.3.1
idna==3.10
charset-normalizer==3.4.3
urllib3==2.5.0

# Git Integration
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2

# Template Engine
Jinja2==3.1.6
MarkupSafe==3.0.2

# Data Serialization
protobuf==6.32.1
pyarrow==21.0.0

# Type Checking and Validation
typing_extensions==4.15.0
typing-inspection==0.4.1
annotated-types==0.7.0

# Web Framework Dependencies
tornado==6.5.2
blinker==1.9.0
cachetools==6.2.0
attrs==25.3.0
six==1.17.0
jiter==0.10.0
narwhals==2.5.0